+++ demos/cypress-shared/cypress/e2e/captcha.cy.ts
@@ -20,7 +20,7 @@ import { type Captcha, CaptchaType } from "@prosopo/types";
 import { at } from "@prosopo/util";
 import { checkboxClass, getWidgetElement } from "../support/commands.js";
 
-let captchaType: CaptchaType;
+const baseCaptchaType: CaptchaType = Cypress.env("CAPTCHA_TYPE") || "image";
 
 describe("Captchas", () => {
 	beforeEach(() => {
@@ -32,8 +32,7 @@ describe("Captchas", () => {
 				console.log("Challenge passed");
 			};
 		});
-		captchaType = Cypress.env("CAPTCHA_TYPE") || "image";
-		cy.registerSiteKey(captchaType).then((response) => {
+		cy.registerSiteKey(baseCaptchaType).then((response) => {
 			// Log the response status and body using cy.task()
 			cy.task("log", `Response status: ${response.status}`);
 			cy.task("log", `Response: ${JSON.stringify(response.body)}`);
@@ -88,12 +87,12 @@ describe("Captchas", () => {
 	});
 
 	after(() => {
-		cy.registerSiteKey(CaptchaType.image);
+		cy.registerSiteKey(baseCaptchaType);
 	});
 
 	it("An error is returned if captcha type is set to pow and the wrong captcha type is used in the widget", () => {
-		expect(captchaType).to.not.equal(CaptchaType.pow);
-		cy.registerSiteKey(CaptchaType.pow).then((response) => {
+		expect(baseCaptchaType).to.not.equal(CaptchaType.pow);
+		cy.registerSiteKey(baseCaptchaType, CaptchaType.pow).then((response) => {
 			// Log the response status and body using cy.task()
 			cy.task("log", `Response status: ${response.status}`);
 			cy.task("log", `Response: ${JSON.stringify(response.body)}`);