+++ packages/provider/src/tasks/powCaptcha/powTasksUtils.ts
@@ -14,8 +14,8 @@
 
 import { sha256 } from "@noble/hashes/sha256";
 import { stringToHex } from "@polkadot/util";
-import { signatureVerify } from "@polkadot/util-crypto";
 import { ProsopoContractError } from "@prosopo/common";
+import { signatureVerify } from "@prosopo/util-crypto";
 
 export const validateSolution = (
 	nonce: number,
@@ -28,13 +28,13 @@ export const validateSolution = (
 		.startsWith("0".repeat(difficulty));
 
 export const checkPowSignature = (
-	challenge: string,
+	message: string,
 	signature: string,
 	address: string,
 	signatureType?: string,
 ): void => {
 	const signatureVerification = signatureVerify(
-		stringToHex(challenge),
+		stringToHex(message),
 		signature,
 		address,
 	);
@@ -43,6 +43,8 @@ export const checkPowSignature = (
 			context: {
 				ERROR: `Signature is invalid for this message: ${signatureType}`,
 				failedFuncName: checkPowSignature.name,
+				address,
+				message,
 				signature,
 				signatureType,
 			},