+++ packages/api-express-router/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @prosopo/api-express-router
 
+## 3.0.0
+### Major Changes
+
+- 64b5bcd: Access Controls
+
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/common@3.0.0
+  - @prosopo/api-route@2.6.5
+
+## 2.6.4
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/api-route@2.6.4
+  - @prosopo/common@2.7.2
+
+## 2.6.3
+### Patch Changes
+
+  - @prosopo/common@2.7.1
+  - @prosopo/api-route@2.6.3
+
+## 2.6.2
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/common@2.7.0
+  - @prosopo/api-route@2.6.2
+
+## 2.6.1
+
+### Patch Changes
+
+- Updated dependencies [04cc7ee]
+  - @prosopo/common@2.6.1
+  - @prosopo/api-route@2.6.1
+
 ## 2.6.0
 
 ### Minor Changes