+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,185 @@
 # @prosopo/cli
 
+## 3.0.0
+### Major Changes
+
+- 64b5bcd: Access Controls
+
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/api-express-router@3.0.0
+  - @prosopo/user-access-policy@3.0.0
+  - @prosopo/provider@3.0.0
+  - @prosopo/common@3.0.0
+  - @prosopo/dotenv@3.0.0
+  - @prosopo/locale@3.0.0
+  - @prosopo/types@3.0.0
+  - @prosopo/util@3.0.0
+  - @prosopo/api@3.0.0
+  - @prosopo/env@3.0.0
+  - @prosopo/config@3.0.0
+  - @prosopo/keyring@2.8.3
+
+## 2.8.0
+### Minor Changes
+
+- aee3efe: Add healthz endpoint
+
+### Patch Changes
+
+- Updated dependencies [aee3efe]
+  - @prosopo/provider@2.14.0
+  - @prosopo/types@2.10.0
+  - @prosopo/api@2.7.2
+  - @prosopo/env@2.7.3
+  - @prosopo/keyring@2.8.2
+  - @prosopo/user-access-policy@2.6.8
+
+## 2.7.4
+### Patch Changes
+
+- Updated dependencies [d5f2e95]
+  - @prosopo/provider@2.13.0
+
+## 2.7.3
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/api-express-router@2.6.4
+  - @prosopo/user-access-policy@2.6.7
+  - @prosopo/util-crypto@13.5.1
+  - @prosopo/provider@2.12.1
+  - @prosopo/keyring@2.8.1
+  - @prosopo/common@2.7.2
+  - @prosopo/dotenv@2.6.4
+  - @prosopo/types@2.9.1
+  - @prosopo/api@2.7.1
+  - @prosopo/env@2.7.2
+  - @prosopo/config@2.6.1
+  - @prosopo/util@2.6.1
+
+## 2.7.2
+### Patch Changes
+
+- Updated dependencies [d6de900]
+  - @prosopo/provider@2.12.0
+  - @prosopo/api@2.7.0
+
+## 2.7.1
+### Patch Changes
+
+- Updated dependencies [30bb383]
+  - @prosopo/util-crypto@13.5.0
+  - @prosopo/provider@2.11.0
+  - @prosopo/keyring@2.8.0
+  - @prosopo/types@2.9.0
+  - @prosopo/common@2.7.1
+  - @prosopo/api@2.6.6
+  - @prosopo/env@2.7.1
+  - @prosopo/user-access-policy@2.6.6
+  - @prosopo/api-express-router@2.6.3
+  - @prosopo/dotenv@2.6.3
+
+## 2.7.0
+### Minor Changes
+
+- 8f0644a: Taking required functions from polkadot/keyring and polkadot/util-crypto in-house and removing WASM dependencies. Adding @scure JS-based sr25519 function instead.
+
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/util-crypto@13.4.0
+  - @prosopo/provider@2.10.0
+  - @prosopo/keyring@2.7.0
+  - @prosopo/common@2.7.0
+  - @prosopo/types@2.8.0
+  - @prosopo/env@2.7.0
+  - @prosopo/api-express-router@2.6.2
+  - @prosopo/dotenv@2.6.2
+  - @prosopo/user-access-policy@2.6.5
+  - @prosopo/api@2.6.5
+
+## 2.6.12
+
+### Patch Changes
+
+- @prosopo/provider@2.9.8
+
+## 2.6.11
+
+### Patch Changes
+
+- @prosopo/provider@2.9.7
+
+## 2.6.10
+
+### Patch Changes
+
+- @prosopo/provider@2.9.6
+
+## 2.6.9
+
+### Patch Changes
+
+- @prosopo/provider@2.9.5
+
+## 2.6.8
+
+### Patch Changes
+
+- Updated dependencies [04cc7ee]
+  - @prosopo/common@2.6.1
+  - @prosopo/api-express-router@2.6.1
+  - @prosopo/dotenv@2.6.1
+  - @prosopo/env@2.6.5
+  - @prosopo/keyring@2.6.4
+  - @prosopo/provider@2.9.4
+  - @prosopo/types@2.7.1
+  - @prosopo/user-access-policy@2.6.4
+  - @prosopo/api@2.6.4
+
+## 2.6.7
+
+### Patch Changes
+
+- @prosopo/provider@2.9.3
+
+## 2.6.6
+
+### Patch Changes
+
+- @prosopo/provider@2.9.2
+
+## 2.6.5
+
+### Patch Changes
+
+- Updated dependencies [5656b0c]
+- Updated dependencies [5656b0c]
+  - @prosopo/provider@2.9.1
+
+## 2.6.4
+
+### Patch Changes
+
+- Updated dependencies [6e1aef6]
+  - @prosopo/provider@2.9.0
+  - @prosopo/types@2.7.0
+  - @prosopo/api@2.6.3
+  - @prosopo/env@2.6.4
+  - @prosopo/keyring@2.6.3
+  - @prosopo/user-access-policy@2.6.3
+
+## 2.6.3
+
+### Patch Changes
+
+- Updated dependencies [cf59998]
+  - @prosopo/provider@2.8.0
+  - @prosopo/env@2.6.3
+
 ## 2.6.2
 
 ### Patch Changes