+++ packages/detector/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prosopo/detector
 
+## 3.0.0
+### Major Changes
+
+- 64b5bcd: Access Controls
+
+## 2.6.1
+
+### Patch Changes
+
+- 5656b0c: Adding cypress tests for invisible
+- 5656b0c: Adding all client examples to bundle example
+
 ## 2.6.0
 
 ### Minor Changes