+++ packages/procaptcha-react/CHANGELOG.md
@@ -1,5 +1,153 @@
 # @prosopo/procaptcha-react
 
+## 2.6.19
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/common@3.0.0
+  - @prosopo/locale@3.0.0
+  - @prosopo/types@3.0.0
+  - @prosopo/util@3.0.0
+  - @prosopo/procaptcha@2.7.5
+  - @prosopo/procaptcha-common@2.7.4
+
+## 2.6.18
+### Patch Changes
+
+- Updated dependencies [aee3efe]
+  - @prosopo/types@2.10.0
+  - @prosopo/procaptcha@2.7.4
+  - @prosopo/procaptcha-common@2.7.3
+
+## 2.6.17
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/procaptcha-common@2.7.2
+  - @prosopo/procaptcha@2.7.3
+  - @prosopo/common@2.7.2
+  - @prosopo/types@2.9.1
+  - @prosopo/util@2.6.1
+
+## 2.6.16
+### Patch Changes
+
+  - @prosopo/procaptcha@2.7.2
+
+## 2.6.15
+### Patch Changes
+
+- Updated dependencies [30bb383]
+  - @prosopo/types@2.9.0
+  - @prosopo/common@2.7.1
+  - @prosopo/procaptcha@2.7.1
+  - @prosopo/procaptcha-common@2.7.1
+
+## 2.6.14
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/procaptcha-common@2.7.0
+  - @prosopo/procaptcha@2.7.0
+  - @prosopo/common@2.7.0
+  - @prosopo/types@2.8.0
+
+## 2.6.13
+### Patch Changes
+
+- cf26d7e: Prevents the translation key (WIDGET.I_AM_HUMAN) to be shown until the translation is loaded
+
+## 2.6.12
+
+### Patch Changes
+
+- ea38a1c: lint
+- Updated dependencies [ea38a1c]
+  - @prosopo/procaptcha-common@2.6.11
+  - @prosopo/procaptcha@2.6.12
+
+## 2.6.11
+
+### Patch Changes
+
+- b2ae723: lint
+- Updated dependencies [b2ae723]
+  - @prosopo/procaptcha-common@2.6.10
+  - @prosopo/procaptcha@2.6.11
+
+## 2.6.10
+
+### Patch Changes
+
+- d17c67f: lint
+- Updated dependencies [d17c67f]
+  - @prosopo/procaptcha-common@2.6.9
+  - @prosopo/procaptcha@2.6.10
+
+## 2.6.9
+
+### Patch Changes
+
+- 0d194f2: lint
+- Updated dependencies [0d194f2]
+  - @prosopo/procaptcha-common@2.6.8
+  - @prosopo/procaptcha@2.6.9
+
+## 2.6.8
+
+### Patch Changes
+
+- Updated dependencies [04cc7ee]
+  - @prosopo/common@2.6.1
+  - @prosopo/procaptcha@2.6.8
+  - @prosopo/procaptcha-common@2.6.7
+  - @prosopo/types@2.7.1
+
+## 2.6.7
+
+### Patch Changes
+
+- bc892fa: lint
+- Updated dependencies [bc892fa]
+  - @prosopo/procaptcha-common@2.6.6
+  - @prosopo/procaptcha@2.6.7
+
+## 2.6.6
+
+### Patch Changes
+
+- 84fc39f: lint
+- Updated dependencies [84fc39f]
+  - @prosopo/procaptcha-common@2.6.5
+  - @prosopo/procaptcha@2.6.6
+
+## 2.6.5
+
+### Patch Changes
+
+- 5656b0c: Adding cypress tests for invisible
+- 5656b0c: Adding all client examples to bundle example
+- Updated dependencies [5656b0c]
+- Updated dependencies [5656b0c]
+  - @prosopo/procaptcha-common@2.6.4
+  - @prosopo/procaptcha@2.6.5
+
+## 2.6.4
+
+### Patch Changes
+
+- Updated dependencies [6e1aef6]
+  - @prosopo/types@2.7.0
+  - @prosopo/procaptcha@2.6.4
+  - @prosopo/procaptcha-common@2.6.3
+
+## 2.6.3
+
+### Patch Changes
+
+- @prosopo/procaptcha@2.6.3
+
 ## 2.6.2
 
 ### Patch Changes