+++ packages/types-database/CHANGELOG.md
@@ -1,5 +1,72 @@
 # @prosopo/types-database
 
+## 3.0.0
+### Major Changes
+
+- 64b5bcd: Access Controls
+
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/user-access-policy@3.0.0
+  - @prosopo/common@3.0.0
+  - @prosopo/types@3.0.0
+
+## 2.7.6
+### Patch Changes
+
+- Updated dependencies [aee3efe]
+  - @prosopo/types@2.10.0
+  - @prosopo/user-access-policy@2.6.8
+
+## 2.7.5
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/user-access-policy@2.6.7
+  - @prosopo/common@2.7.2
+  - @prosopo/types@2.9.1
+
+## 2.7.4
+### Patch Changes
+
+- Updated dependencies [30bb383]
+  - @prosopo/types@2.9.0
+  - @prosopo/common@2.7.1
+  - @prosopo/user-access-policy@2.6.6
+
+## 2.7.3
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/common@2.7.0
+  - @prosopo/types@2.8.0
+  - @prosopo/user-access-policy@2.6.5
+
+## 2.7.2
+
+### Patch Changes
+
+- Updated dependencies [04cc7ee]
+  - @prosopo/common@2.6.1
+  - @prosopo/types@2.7.1
+  - @prosopo/user-access-policy@2.6.4
+
+## 2.7.1
+
+### Patch Changes
+
+- Updated dependencies [6e1aef6]
+  - @prosopo/types@2.7.0
+  - @prosopo/user-access-policy@2.6.3
+
+## 2.7.0
+
+### Minor Changes
+
+- cf59998: Update DB schema
+
 ## 2.6.2
 
 ### Patch Changes